Refactor discount curve calculations across various instruments
- Updated discount curve calculations in multiple test files to remove unnecessary type casting, enhancing code clarity and consistency.
- Standardized the use of `exp()` function across various instruments, ensuring uniformity in discount factor calculations.
- Improved test cases for various financial instruments, including bonds, swaps, and options, to reflect the updated calculations and enhance accuracy in results.
- Removed redundant code and improved documentation in test files to streamline the testing process and improve maintainability.

+++ finstack/valuations/tests/instruments/bond/pricing.rs
@@ -21,9 +21,9 @@ fn create_flat_curve(rate: f64, base_date: Date, id: &str) -> DiscountCurve {
         .day_count(DayCount::Act365F)
         .knots([
             (0.0, 1.0),
-            (1.0, (-rate).exp() as f64),
-            (5.0, (-rate * 5.0).exp() as f64),
-            (10.0, (-rate * 10.0).exp() as f64),
+            (1.0, (-rate).exp()),
+            (5.0, (-rate * 5.0).exp()),
+            (10.0, (-rate * 10.0).exp()),
         ])
         .build()
         .unwrap()
