+++ finstack/valuations/tests/instruments/cds/test_cds_metrics.rs
@@ -19,9 +19,9 @@ fn build_test_discount(rate: f64, base: Date, id: &str) -> DiscountCurve {
         .day_count(DayCount::Act360)
         .knots([
             (0.0, 1.0),
-            (1.0, (-rate).exp() as f64),
-            (5.0, (-rate * 5.0).exp() as f64),
-            (10.0, (-rate * 10.0).exp() as f64),
+            (1.0, (-rate).exp()),
+            (5.0, (-rate * 5.0).exp()),
+            (10.0, (-rate * 10.0).exp()),
         ])
         .build()
         .unwrap()
