+++ finstack/valuations/tests/instruments/common/models.disabled/test_trinomial_tree.rs
@@ -380,7 +380,7 @@ fn test_put_call_parity() {
 
     // Assert: C - P = S - K*e^(-rT)
     let lhs = call_price - put_price;
-    let rhs = spot - strike * (-rate * time).exp() as f64;
+    let rhs = spot - strike * (-rate * time).exp();
 
     assert_approx_eq(lhs, rhs, 0.1, "Put-call parity");
 }
@@ -897,7 +897,7 @@ fn test_low_volatility() {
 
     // Assert: Should be close to discounted intrinsic
     let intrinsic = 5.0;
-    let discounted = intrinsic * (-0.05 * 1.0).exp() as f64;
+    let discounted = intrinsic * (-0.05 * 1.0).exp();
     assert_relative_eq(price, discounted, 0.1, "Low vol ~ discounted intrinsic");
 }
 
@@ -938,7 +938,7 @@ fn test_zero_strike_call() {
     let price = tree.price(vars, time, &ctx, &valuator).unwrap();
 
     // Assert: Should equal discounted spot
-    let expected = spot * ((-rate * time).exp() as f64);
+    let expected = spot * ((-rate * time).exp());
     assert_relative_eq(price, expected, 0.01, "Zero strike call ~ discounted spot");
 }
