+++ finstack/valuations/tests/instruments/irs/metrics/bucketed_dv01.rs
@@ -16,12 +16,12 @@ fn build_flat_discount_curve(rate: f64, base_date: Date) -> DiscountCurve {
         .day_count(DayCount::Act360)
         .knots([
             (0.0, 1.0),
-            (1.0, (-rate).exp() as f64),
-            (2.0, (-rate * 2.0).exp() as f64),
-            (3.0, (-rate * 3.0).exp() as f64),
-            (5.0, (-rate * 5.0).exp() as f64),
-            (7.0, (-rate * 7.0).exp() as f64),
-            (10.0, (-rate * 10.0).exp() as f64),
+            (1.0, (-rate).exp()),
+            (2.0, (-rate * 2.0).exp()),
+            (3.0, (-rate * 3.0).exp()),
+            (5.0, (-rate * 5.0).exp()),
+            (7.0, (-rate * 7.0).exp()),
+            (10.0, (-rate * 10.0).exp()),
         ])
         .build()
         .unwrap()
