+++ finstack/valuations/tests/instruments/irs/metrics/par_rate.rs
@@ -18,9 +18,9 @@ fn build_flat_curves(rate: f64, base_date: Date) -> MarketContext {
         .day_count(DayCount::Act360)
         .knots([
             (0.0, 1.0),
-            (1.0, (-rate).exp() as f64),
-            (5.0, (-rate * 5.0).exp() as f64),
-            (10.0, (-rate * 10.0).exp() as f64),
+            (1.0, (-rate).exp()),
+            (5.0, (-rate * 5.0).exp()),
+            (10.0, (-rate * 10.0).exp()),
         ])
         .build()
         .unwrap();
