+++ finstack/valuations/tests/instruments/irs/metrics/pv_fixed.rs
@@ -16,9 +16,9 @@ fn build_flat_discount_curve(rate: f64, base_date: Date) -> DiscountCurve {
         .day_count(DayCount::Act360)
         .knots([
             (0.0, 1.0),
-            (1.0, (-rate).exp() as f64),
-            (5.0, (-rate * 5.0).exp() as f64),
-            (10.0, (-rate * 10.0).exp() as f64),
+            (1.0, (-rate).exp()),
+            (5.0, (-rate * 5.0).exp()),
+            (10.0, (-rate * 10.0).exp()),
         ])
         .build()
         .unwrap()
